refactored elitism
